backend/requirements.txt
Flask>=2.0
flask-cors
flask-jwt-extended
pymongo
cryptography
bcrypt
pyotp
qrcode
pillow
python-dotenv